+++ Tests/UnitTests/Views/ScrollBarTests.cs
@@ -283,11 +283,10 @@ public class ScrollBarTests (ITestOutputHelper output)
         Assert.Equal (1, count);
     }
 
-    [Theory]
-    [SetupFakeDriver]
-
     #region Draw
 
+    [Theory]
+    [SetupFakeDriver]
 
     #region Horizontal
