+++ Tests/UnitTestsParallelizable/Views/NumericUpDownTests.cs
